Fix bugs with dict / set "deleted" values
If we tried to remove a dict item, it wouldn't take into account whether that item had already been deleted. This means it could potentially try to delete the same key twice. Inserting into a dict also checked if a key was "deleted". This means you could insert into a deleted key and end up with two of the same key within the dict