feat: add list.sortFunc
Adds a function to list implemented in dictu which uses quickSort
to allow custom sorting which is suuuper useful for a lot of problems.
I would have preferred to implement this is C too but there doesn't seam
to be a straight forward way to pass callbacks into c.
Using a hybrid approach would have been hacky too, but of course this approach
does duplicate the sorting logic.