fix: when executing a child process, do not waitpid before reading
Trying to waitpid before reading the piped content will lead to a blocked pipe and
get the child process "stuck". So it should be read BEFORE waitpid is called.
Further there was a bug within the reallocation logic leading to heap corruption
because the comparison size did not include the just read chunk, leading to heap corruption,
which itself led to reallocate failiing.