feat: call dictu from native code
This uses a stack loaded piece of code to interrupt the current
vm execution in order to invoke a function called by native code.

It creates a new callframe for that and then makes sure the vm loop
breaks after that frame was reached again, this way the result is
returned tp the caller. It then resets the vm state and continues
normally.